refactor: extract server context logic to file
